requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
openai
beautifulsoup4
requests
ddgs # duckduckgo-search